requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Werkzeug==2.2.2
apscheduler==3.10.0
google-generativeai==0.1.0
pytz==2023.3
requests==2.28.1
pandas==1.5.3
